requirements/base.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
psycopg==3.3.1
psycopg-binary==3.3.1
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
requests==2.32.5
rich==14.2.0
sqlparse==0.5.4
urllib3==2.6.2
whitenoise==6.11.0
